refactor: group loose components into role folders
Follows the existing convention (dock/, topbar/, layout/): one folder per role
with its desktop + mobile variants together.

  window/   AppWindow, WindowChrome, WindowTopbar, WindowsContent, FinderItems
  body/     BodyContent, MobileBodyContent
  widgets/  DesktopWidgets, MobileWidgets
  launcher/ DesktopIcons, MobileAppMenu

getIcon, Clock, Maintenance and NotSupport stay at the root: the first two are
used everywhere (root signals that), the last two are single-use fallback
screens with no pair partner. No catch-all common/ folder, which would just
collect whatever lands next.
