+++ src/components/index.ts
@@ -1,8 +1,8 @@
 export { default as Clock } from "./Clock";
 export { default as Maintenance } from "./Maintenance";
-export { default as BodyContent } from "./BodyContent";
-export { default as MobileAppMenu } from "./MobileAppMenu";
-export { default as MobileBodyContent } from "./MobileBodyContent";
 export { default as NotSupport } from "./NotSupport";
-export { default as WindowsContent } from "./WindowsContent";
-export { default as WindowTopbar } from "./WindowTopbar";
+export { default as BodyContent } from "./body/BodyContent";
+export { default as MobileBodyContent } from "./body/MobileBodyContent";
+export { default as MobileAppMenu } from "./launcher/MobileAppMenu";
+export { default as WindowsContent } from "./window/WindowsContent";
+export { default as WindowTopbar } from "./window/WindowTopbar";
