+++ src/components/layout/DesktopLayout.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import useScreenSize from "../../utils/useScreenSize";
-import BodyContent from "../BodyContent";
-import DesktopIcons from "../DesktopIcons";
-import DesktopWidgets from "../DesktopWidgets";
+import BodyContent from "../body/BodyContent";
+import DesktopIcons from "../launcher/DesktopIcons";
+import DesktopWidgets from "../widgets/DesktopWidgets";
 import { DesktopTopBar } from "../topbar";
 import { DockMenu } from "../dock";
