+++ src/components/layout/MobileLayout.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import useScreenSize from "../../utils/useScreenSize";
 import { MobileTopBar } from "../topbar";
 import { MobileDockMenu } from "../dock";
-import MobileBodyContent from "../MobileBodyContent";
+import MobileBodyContent from "../body/MobileBodyContent";
 
 const MobileLayout: React.FC = () => {
   const { width, height } = useScreenSize();
