+++ src/components/window/FinderItems.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "../utils/cn";
+import { cn } from "../../utils/cn";
 
 type FinderGridItemProps = {
   label: string;
