+++ src/views/window/desktop/CalculatorWindow.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AppWindow from "../../../components/AppWindow";
+import AppWindow from "../../../components/window/AppWindow";
 import { Calculator } from "../../../components/calculator";
 
 const CalculatorWindow: React.FC = () => (
