+++ src/views/window/desktop/CalendarWindow.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AppWindow from "../../../components/AppWindow";
+import AppWindow from "../../../components/window/AppWindow";
 import Calendar from "../../../components/calendar/Calendar";
 
 const CalendarWindow: React.FC = () => (
