+++ src/views/window/desktop/ProjectsWindow.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import AppWindow from "../../../components/AppWindow";
-import { WindowBody } from "../../../components/WindowChrome";
+import AppWindow from "../../../components/window/AppWindow";
+import { WindowBody } from "../../../components/window/WindowChrome";
 import { ProjectIndexType, ViewMode } from "../../../types";
 import ProjectsContent from "../../../components/projects/ProjectsContent";
 import useNavHistory from "../../../utils/useNavHistory";
