+++ src/views/window/desktop/TerminalWindow.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AppWindow from "../../../components/AppWindow";
+import AppWindow from "../../../components/window/AppWindow";
 import Terminal from "../../../components/terminal/Terminal";
 
 const TerminalWindow: React.FC = () => (
